GP-47112 Fix dynamic action provider class generation
This changes the approach used to generate dynamic action provider
classes for individual SQL Tasks. The previous approach using
anonymous classes and class_alias() did not behave as expected,
since anonymous classes in PHP are ... special and appear to be just
syntactic sugar. Static variables are shared even if the class
is defined (or aliased) multiple times.

There does not appear to be any other way to get to the task ID, so
this uses an ugly hack where the class code is generated via smarty
and then eval'd.